refactor(e2e): deduplicate BDD steps and reuse Navigation class
- Move shared navigation steps to tests/ui/steps/navigation.ts using Navigation.goToSidebar()
- Remove waitForLoadState(networkidle) per reviewer feedback
- Remove duplicate step definitions from @search/search.step.ts
- Delete @page-title/page-title.step.ts (steps moved to shared location)
- Create @search/search.feature for TC-3248 upload advisory button test
- Remove duplicate scenario from search.feature
- Update verify-page-title.feature examples to match sidebar menu names
