--- e2e/tests/ui/features/@page-title/page-title.step.ts
@@ -1,3 +1,0 @@
-// Import page-level common steps - no need to define steps here
-// All steps used by @page-title features are defined in page-common-steps.ts
-import "../page-common-steps";
