+++ e2e/tests/ui/features/@page-title/verify-page-title.feature
@@ -4,22 +4,7 @@ Feature: Page Title Updates
 	So that I can distinguish between multiple tabs and use browser history effectively
 
 # Related to TC-3370: Page title does not change based on viewed page
-# The fix ensures page titles update when navigating to different pages
-Scenario: Verify page title changes when navigating to Search page
-	When User navigates to Search results page
-	Then the page title should contain "Search"
-
-# Note: Tabs within the Search page all show "Search" in the title
-# because you're still on the Search page - just viewing different result types
-Scenario: Verify page title remains "Search" when switching between tabs
-	When User navigates to Search results page
-	Then the page title should contain "Search"
-	When Clicks on Advisories tab
-	Then the page title should contain "Search"
-	When User clicks on "SBOMs" tab
-	Then the page title should contain "Search"
-	When User clicks on "Packages" tab
-	Then the page title should contain "Search"
+# The ensures page titles update when navigating to different pages
 
 # Verify page titles for main application pages
 Scenario Outline: Verify page title changes when navigating to <page> page
@@ -28,8 +13,24 @@ Scenario Outline: Verify page title changes when navigating to <page> page
 
 	Examples:
 		| page            | expectedTitle   |
-		| SBOMs           | SBOMs           |
+		| Dashboard       | Home            |
+		| Search          | Search          |
+		| All SBOMs       | SBOMs           |
 		| Advisories      | Advisories      |
 		| Vulnerabilities | Vulnerabilities |
 		| Packages        | Packages        |
 		| Importers       | Importers       |
+
+# Note: Tabs within the Search page all show "Search" in the title
+# because you're still on the Search page - just viewing different result types
+Scenario Outline: Verify page title remains "Search" when switching to <tabName> tab
+	When User navigates to "Search" page
+	And User clicks on "<tabName>" tab
+	Then the page title should contain "Search"
+
+	Examples:
+		| tabName         |
+		| Advisories      |
+		| SBOMs           |
+		| Packages        |
+		| Vulnerabilities |
